+++ NpgsqlRestTests/TsClientTests/GetConfigTests.cs
@@ -94,7 +94,7 @@ export async function tsclientTestGetConfigStatus() : Promise<{status: number, r
     return {
         status: response.status,
         response: response.ok ? await response.json() : undefined!,
-        error: !response.ok ? await response.json() as {status: number; title: string; detail?: string | null} : undefined
+        error: !response.ok && response.headers.get("content-length") !== "0" ? await response.json() as {status: number; title: string; detail?: string | null} : undefined
     };
 }