+++ NpgsqlRestTests/TsClientTests/GetUserDataTests.cs
@@ -142,7 +142,7 @@ export async function tsclientTestGetUserDataStatus() : Promise<{status: number,
     return {
         status: response.status,
         response: response.ok ? await response.json() as ITsclientTestGetUserDataStatusResponse[] : undefined!,
-        error: !response.ok ? await response.json() as {status: number; title: string; detail?: string | null} : undefined
+        error: !response.ok && response.headers.get("content-length") !== "0" ? await response.json() as {status: number; title: string; detail?: string | null} : undefined
     };
 }